--- src/UI/LoginButton/LoginButtonSSR.tsx
@@ -1,6 +1,0 @@
-import { withStyles } from 'hocs/withStyles';
-import { LoginButtonComponent } from './LoginButtonComponent';
-
-export const LoginButtonSSR = withStyles(LoginButtonComponent, {
-  local: () => import('UI/LoginButton/loginButtonStyles.scss')
-});
