+++ src/UI/extension/ExtensionLoginButton/index.ts
@@ -1,2 +1,1 @@
 export * from './ExtensionLoginButton';
-export * from './types';
